[interpreter] implement fee accounting
This does not currently implement stack accounting directly as specified
in the specification. The reason for that is two-fold – first is that
the stack in the interpreter is split between values and
frames/labels/instructions. The second is that the stack is represented
as a basic list/vector data structure, so there isn’t a convenient
hook-point for push or pop operations.