spec: adjust when the stack limits are evaluated
This changes the specification to require that the stack limits are
enforced at the function call, rather than on every stack operation.

Doing the latter requires that, in case a `trap` does occur, any side
effects are executed before the `trap`. However, in practice many VM
implementations will want to allocate the stack space upon the function
entry and the way the reserved stack space is used is often
unpredictable.

As thus there isn’t a guarantee that the more-granularly enforced stack
validation will actually interact correctly with the realities of the
implementation concerns.

Unfortunately this also invalidates the implementation in the reference
interpreter. The current reference will output the stack size trace
throughout the execution, however given a snippet along the lines of

```wast
if
    (i32.const 42)
    drop
else
    (v128.const 42)
    drop
end
```

the interpreter would determine the maximum stack usage to be 4 bytes
worth of values if the `if` branch was taken. The typical VM
implementation, in absence of any other optimizations, would be required
to reserve at least 16 bytes instead to also account for the `else`
branch.