spec: describe how the max stack size is computed
Admittedly I wasn’t in a mood to see if I can write this in a way that's
very spec-esque, and so the language in this appendix will read quite
weird, and probably leaves a bunch of space for interpretations too.

Should be good enough for the initial intent though.