max_stack: review and refactor slightly
This commit is a result of a thorough review of all instructions. In
order to aid this effort, I added some macros that allowed to compress
and bring instruction definitions somewhat closer to those in the
specification. There are still cases where ad-hoc implementations were
required (most notably for control instructions) and in some cases I
chose to unify the categories further than the spec did (e.g.
v*binop delegate to `visit_binop`, despite not being formally a part of
the `binop` category in the spec.)

Furthermore, I also adjusted the code to use only a single operand
stack, rather than multiple sub-stacks per-frame. This simplified some
parts of the code, and should be a fair bit faster (single heap vector
to maintain, rather than many.) After such adjustment, the
implementation of this analysis is also quite a bit more resembling of
the reference validator implementation as well.