max_stack: non-functional style refactors
For the delegated visit methods specify the return value as a type
alias. This is going to be easier to modify in the future and less stuff
to read through now.

Stop threading through the `offset` through the arguments all over the
place. It is a lot of noise and `pop(offset)` just reads weird. Since
this is all methods on a `Visitor` implementation already, we can just
stash the `offset` as a field instead.