max_stack: Fix handling of block parameters
Previous implementation would fail to pop parameters from the parent
operand stack before pushing the frame and parameters. This would
effectively result in incorrect operand stack with too many values on
it.

The fact that this was not caught earlier is an unfortunate consequence
of the fact that we're not actually tracking the types being popped in
our implementation, so there isn't a good way to cross-check for this.

Ideally we get a nice reference to cross-check with somehow, but that
also has proven to be difficult...?