gas: add an implementation of gas analysis 1st pass
This analysis walks the instruction stream, computing a table of
instruction ranges, the associated costs and kinds. In the second pass
we will compartmentalize/optimize the table by merging adjacent table
entries based on a ruleset (such as, two pure instructions can be merged
together) which would reduce the amount of gas instrumentation points
that need to be inserted.

Note, that the correctness of the first pass is quite important for a
correct overall result. In large part the first pass is responsible for
annotating all points in the program across which the gas accounting
cannot be aggregated together. Examples of these points range anywhere
from various branches to all sorts of side effects and aggregate
instructions.