gas: use instruction categories for partial insns
Quite a few WebAssembly instructions are partial and may trap based on
the operands it needs to handle. Notable examples of such instructions
are the instructions that interact with the memory, or those that may
attempt an integer divide by 0.

Since evaluation of these instructions may trap, we must make sure that
the gas counter (inclusive of the potentially trapping instruction) is
exact before the instruction in question is executed.