test: insert instrumentation functions and types
As we insert the imports for our instrumentation functions, the index
space of the functions shifts by two. This means that every reference to
a function also needs to shift by two. The intended diff for the
snapshots is basically just the same – function indices shift, but
everything else stays the same.