test: instrument with `$finite_wasm_stack`
We add a function call to the first of our instrumentation functions at
the beginning of each function. This will inform the embedder that the
function is going to use the specified amount of stack space so that the
interpreter can act accordingly, possibly enforcing a stack limit of
some sort.

Note that this is not sufficient for most production use-cases as there
is no paired operation to reduce the amount of stack used when the
function returns!