test: instrument gas
Similarly, to the previous commit, this time for the gas. We introduce
calls to the `$finite_wasm_gas` function so that the interpreter can act
on it and output information that we would later cross-reference with
the interpreter's own execution log. That way we can validate that the
instrumentation is actually doing what we expect it to do.