interpreter: print on gas/stack instrumentations
This prints out some information each time gas/stack instrumentations
are invoked. As the interpreter is not entirely aware of these
intrinsics being free to invoke from the basic appearance of the
instruction sequence, we mitigate this by applying a `-2` cost to
invocation of these intrinsics, effectively undoing the charges that
have occurred until that point.