interpreter: remove stack tracing support
This is O(n^2) and makes the test suite execute for dozens of minutes,
which is unacceptable. Not to mention that it we don’t know of a good
way to actually utilize data from the interpreter here. It seems like
it'd be more feasible to manually spot-check test cases or at least
approach validating this in some other way.