test: remove selects which wast handles incorrectly
wasmparser models this instruction as `visit_typed_select` with a result
forced to be arity = 1 during decoding. This is a slight divergence from
the interpreter and the spec, which instead rejects the wrong arity
during validation.

The actual problem comes up in `wast` which instead of combining all of
the `(result)` annotations together to end up with arity = 1 for all
these tests, sometimes ends up with an empty list of results.