test: don't bother diffing if updating snapshots
I have seen the diffing code take a very long time to come up with a
differential for some of the changes, to the point where the test suite
felt stuck. We, however, don’t need to diff at all when the user has
requested to update the snapshots anyway. Save some compute effort.