gas: fix handling of the `loop` instruction
Specification-wise the `loop` instruction is evaluated on every
iteration, so we must make sure the gas fee for this instruction gets
applied in such a manner as well. This is a little bit harder than it
sounds, however – a `br` targetting the frame of `loop` will start
executing from the `loop` instruction onwards, so we don't have a good
opportunity to introduce an instrumentation point that _precedes_ this
instruction. Instead we're sort-of forced to bite the bitter pill and
move this instrumentation point to after the `loop` instruction, so that
it is evaluated on every iteration.

Interestingly this might not be as big of an issue for users of this
analysis who aren't instrumenting wasm->wasm code. For them it might not
be a problem to move the instrumentation point to the beginning of the
iteration body. In practice they probably don't need to think about as
`loop` codegens to nothing typically, but it is something to keep in
mind when thinking about future improvements we could make to the way
analysis results are represented.