test: add a self-test for interpreter's intrinsic gas accounting
The interpreter applies a negative gas fee whenever it encounters an
instrumentation intrinsic. `-3` for the `finite_wasm_stack` intrinsic
and `-2` for the `finite_wasm_gas` intrinsic. These correspond to the 3
and 2 gas consumed to set up the arugments for, and to invoke the
intrinsics call in question. Net result is that after executing the
sequence the final gas cost, as observed by the interpreter is 0.

These tests verify this property, making sure that we aren't
accidentally subtracting too much or too little gas.