analysis: improve doc-comments
gas: Looking at the output of rustdoc there were a couple of obvious format
issues. Ideally these tables would have become proper HTML tables. I’m not
willing to sacrifice the source-code readability for nice tables in HTML so
wrapping those in `<pre>` so that formatting is retained is going to be the
solution.

stack: it inherited documentation from the old original crate-level
documentation which was now wrong.