analysis: refactor the public API
With this refactor we now pick names that are overall more sensible, and
the structure is easier to understand at a glance, as it no longer
relies on traits defined in other crates as much. We still want
`gas::CostModel` implementers to also implement a `VisitOperator` but it
is now easier to direct and explain how to do so.