instrument: support for releasing stack in instrumented code
The approach taken here is to wrap-around the function body into
an appropriately typed `block...end` sequence. That way any execution
path that would normally have reached the exit from the function
(including various `*_br`anches  targetting the function frame) now exit
this block instead, giving us an opportunity to release the stack
allocation.

Some instructions that skip this mechanism such as `return_call` and
`return` get instrumented additionally. There are few enough of these
that it wouldn't bother me too much, though.

Various traps and side effects are not handled at all. If the function
executes something like `unreachable`, the stack size will not get reset
to 0, it is a responsibility for the VM to reset the limits after such
an event by themselves. We could implement an alternative where we call
`finite_wasm_stack` with some magic value arguments, or even introduce a
separate function to release all the stack before a trap, but that would
only handle the most basic cases where we know a trap _will_ occur. It
wouldn't help at all for “potentially traps” kinds of instructions, as
these continue using stack during execution (so we can’t release it all)
and might just return successfully, in which case we would have to
re-reeserve the stack space…