max_stack: move buffer lifetime from GAT to trait
This makes the `max_stack::Config` the same as `gas::Config`. The `'b`
lifetime for the input `wasm` module buffer makes more sense as a
generic on the trait: the `'b` is propagated into the
`VisitOperator`, the lifetime of which is then basically intended to
represent data borrowed out of the input module buffer.