all: add a benchmark comparing finite-wasm with wasm-instrument
This benchmark analyzes and instruments a module, doing very roughly the
same thing. We want to quantify just how much faster `finite-wasm`
really is and this benchmark can act as a reasonable gateway to conduct
this investigation. This infrastructure will also help comparing various
performance cliffs we've encountered with wasm-instrument – we can check
if finite-wasm has fixed those we discovered so far.