test: exclude the snapshots from being packaged and published
These are pretty large and at the same time pretty useless for somebody
looking to obtain the crate. If they want to develop the crate, they
should be fetching the git repository, and even if they must use the
`.crate` they can regenerate these snapshots easily too. No reason to
have every use of this crate result in many megs of download.