gas: move flushing of scheduled instrumentation points to set_offset
This makes sure that every time state advances to the next instruction,
the instrumentation point is immediately flushed, even if the
instruction visited does not introduce any of the instrumentation points
itself for whatever reason.

This then also simplifies checks and resilience against mistakes such as
pushing an instrumentation point after the same instruction twice.