max_stack: return an error for types that are too large
wasmparser uses a compressed representation for reference types, which
means that it is unable to represent all types that are representable by
the input text. This is probably fine, because the too-large-type would
be a `Ref(Func(type_index))` for very large values of `type_index`, and
one would need a huge amount of bytes to even produce that many types in
the first place, so the module is approximately guaranteed to be
somehow malformed in the first place.